Commit 2 Parallax
Made a GM so GameObject used as Backgrounds can scroll depending what the camera can see. The camera now follows the player around. So when the player gets close to the edge of a Background previous Backgrounds move to thr front or back depending where the player is. So there is a smooth transition so it looks like the player is going through space in a loop