Commit 3 Player Restrictions
I edited the base class so the PC derived class could inherit a new function. This being Movement restriction. To make sure the PC or any other GameObject doesnt go out of the play area i restricted the transform. Also adding a new script to the Camera. Whcih allows for the camera to follow the player and also stop at certain Y coordinates. So it gives the effect the player has reached the end of the screen.

Project change. Decided the art for the re make of defender will be cartoon based so Cel Shaded