Commit 4
After thinking about how parallax could make the game good or bad. I decided that screenwrapping had to implemented. Without it The game would break for NPCs being spawned would be left behind and if childed to the sprited background they would be restricted and there is only so much there can be done with difficulty in a fixed location. So for now ive added a basic screenwrapping script for 2 box colliders that act as triggers. However later i will be controlling the screenwrapping through the base class