Abduction Mechanic
At the current time this mechanic works to a degree. The NPC Abducter can stop moving randomly on the screen when a raycast hits the human. The abducter then goes down the Y axis until within 2 away. The simply childs the Human to the abducter ship and moves up. Bad thing is that it needs refinement as it cannot monitor multiple humans. As well as when the human moves away from the NPC there behavouir doesnt change and simple continues to move down. However this can be resolved by reversering some statements so if the NPC abducter is more than 2 away then start moving randomly again